// music-manage/src/enums/router-name.ts
export const enum RouterName {
  Home = "/home",
  Info = "/info",
  Song = "/song",
  Singer = "/singer",
  SongList = "/songSheet",
  ListSong = "/listSong",
  Comment = "/Comment",
  Consumer = "/consumer",
  Collect = "/collect",
  Banner = "/banner",
  Error = "/404",
  SignIn = "/",
  SignOut = "0",
}